requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
rest-framework-roles==1.0.6
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
